feat: move `ParachainInherentData` type to `types` pkg
